Migrate position utils to typescript (#3813)
