+++ components/at_mention/at_mention.jsx
@@ -9,7 +9,7 @@ import {displayUsername} from 'mattermost-redux/utils/user_utils';
 
 import ProfilePopover from 'components/profile_popover';
 
-import {popOverOverlayPosition} from 'utils/position_utils.jsx';
+import {popOverOverlayPosition} from 'utils/position_utils.tsx';
 const spaceRequiredForPopOver = 300;
 
 export default class AtMention extends React.PureComponent {
@@ -49,7 +49,7 @@ export default class AtMention extends React.PureComponent {
 
     handleClick = (e) => {
         const targetBounds = this.overlayRef.current.getBoundingClientRect();
-        const placement = popOverOverlayPosition(targetBounds, window.innerHeight, {above: spaceRequiredForPopOver});
+        const placement = popOverOverlayPosition(targetBounds, window.innerHeight, spaceRequiredForPopOver);
 
         this.setState({target: e.target, show: !this.state.show, placement});
     }
